+++ src/bindings.d.ts
@@ -1,8 +1,8 @@
-export {}
-
 declare global {
     const AUTH0_DOMAIN: string
     const AUTH0_CLIENT_ID: string
     const AUTH0_REDIRECT_URI: string
     const AUTH0_URI: string
 }
+
+export {}